feature: run full path commands
